backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.7
chromadb==0.4.18
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==1.1.0
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
mcp==0.1.0
mcp-client==0.1.0
weaviate-client==4.4.4
pinecone-client==2.2.4
